# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Her Planner API - AI-powered life planner for women
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Profile {
  id: number;
  name: string;
  hasKids: boolean;
  numberOfKids?: number | null;
  workSchedule?: string | null;
  healthConditions?: string | null;
  averageSleepHours?: number | null;
  cycleLength?: number | null;
  periodLength?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProfileBody {
  name: string;
  hasKids: boolean;
  numberOfKids?: number | null;
  workSchedule?: string | null;
  healthConditions?: string | null;
  averageSleepHours?: number | null;
  cycleLength?: number | null;
  periodLength?: number | null;
}

export interface UpdateProfileBody {
  name?: string;
  hasKids?: boolean;
  numberOfKids?: number | null;
  workSchedule?: string | null;
  healthConditions?: string | null;
  averageSleepHours?: number | null;
  cycleLength?: number | null;
  periodLength?: number | null;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type TaskView = (typeof TaskView)[keyof typeof TaskView];

export const TaskView = {
  today: "today",
  week: "week",
  month: "month",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  category: string;
  priority: TaskPriority;
  completed: boolean;
  dueDate?: string | null;
  view: TaskView;
  aiSuggested: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CreateTaskBodyView =
  (typeof CreateTaskBodyView)[keyof typeof CreateTaskBodyView];

export const CreateTaskBodyView = {
  today: "today",
  week: "week",
  month: "month",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string | null;
  category: string;
  priority: CreateTaskBodyPriority;
  dueDate?: string | null;
  view: CreateTaskBodyView;
  aiSuggested?: boolean;
}

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type UpdateTaskBodyView =
  (typeof UpdateTaskBodyView)[keyof typeof UpdateTaskBodyView];

export const UpdateTaskBodyView = {
  today: "today",
  week: "week",
  month: "month",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  category?: string;
  priority?: UpdateTaskBodyPriority;
  completed?: boolean;
  dueDate?: string | null;
  view?: UpdateTaskBodyView;
}

export type TasksSummaryToday = {
  total: number;
  completed: number;
};

export type TasksSummaryWeek = {
  total: number;
  completed: number;
};

export type TasksSummaryMonth = {
  total: number;
  completed: number;
};

export type TasksSummaryByCategoryItem = {
  category: string;
  total: number;
  completed: number;
};

export interface TasksSummary {
  today: TasksSummaryToday;
  week: TasksSummaryWeek;
  month: TasksSummaryMonth;
  byCategory: TasksSummaryByCategoryItem[];
}

export type CycleEntryEntryType =
  (typeof CycleEntryEntryType)[keyof typeof CycleEntryEntryType];

export const CycleEntryEntryType = {
  period_start: "period_start",
  period_end: "period_end",
  ovulation: "ovulation",
  symptom: "symptom",
  note: "note",
} as const;

export interface CycleEntry {
  id: number;
  entryType: CycleEntryEntryType;
  date: string;
  symptoms?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateCycleEntryBodyEntryType =
  (typeof CreateCycleEntryBodyEntryType)[keyof typeof CreateCycleEntryBodyEntryType];

export const CreateCycleEntryBodyEntryType = {
  period_start: "period_start",
  period_end: "period_end",
  ovulation: "ovulation",
  symptom: "symptom",
  note: "note",
} as const;

export interface CreateCycleEntryBody {
  entryType: CreateCycleEntryBodyEntryType;
  date: string;
  symptoms?: string | null;
  notes?: string | null;
}

export type CyclePhasePhase =
  (typeof CyclePhasePhase)[keyof typeof CyclePhasePhase];

export const CyclePhasePhase = {
  menstrual: "menstrual",
  follicular: "follicular",
  ovulation: "ovulation",
  luteal: "luteal",
  unknown: "unknown",
} as const;

export interface CyclePhase {
  phase: CyclePhasePhase;
  dayInCycle?: number | null;
  nextPeriodIn?: number | null;
  energyExpectation: string;
  moodExpectation: string;
  recommendation: string;
}

export interface DailyContext {
  id: number;
  date: string;
  sleepHours?: number | null;
  energyLevel?: number | null;
  mood?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateDailyContextBody {
  date: string;
  sleepHours?: number | null;
  energyLevel?: number | null;
  mood?: string | null;
  notes?: string | null;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListTasksParams = {
  view?: ListTasksView;
  category?: string;
  completed?: boolean;
};

export type ListTasksView = (typeof ListTasksView)[keyof typeof ListTasksView];

export const ListTasksView = {
  today: "today",
  week: "week",
  month: "month",
  all: "all",
} as const;

export type ListCycleEntriesParams = {
  limit?: number;
};

export type ListDailyContextsParams = {
  limit?: number;
};