# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Her Planner API - AI-powered life planner for women
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cycle/entries:
    get:
      operationId: listCycleEntries
      tags: [cycle]
      summary: List cycle tracking entries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Cycle entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CycleEntry"
    post:
      operationId: createCycleEntry
      tags: [cycle]
      summary: Log a cycle entry (period start, ovulation, symptoms)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCycleEntryBody"
      responses:
        "201":
          description: Created cycle entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleEntry"
...
  /daily-context:
    get:
      operationId: listDailyContexts
      tags: [dailyContext]
      summary: List daily context entries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily contexts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyContext"
    post:
      operationId: createDailyContext
      tags: [dailyContext]
      summary: Log today's context (sleep, energy, mood)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyContextBody"
      responses:
        "201":
          description: Created daily context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyContext"
  /daily-context/today:
    get:
      operationId: getTodayContext
      tags: [dailyContext]
      summary: Get today's context entry
      responses:
        "200":
          description: Today's context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyContext"
        "404":
          description: No entry for today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CycleEntry:
      type: object
      properties:
        id:
          type: integer
        entryType:
          type: string
          enum: [period_start, period_end, ovulation, symptom, note]
        date:
          type: string
          format: date
        symptoms:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - entryType
        - date
        - createdAt
    CreateCycleEntryBody:
      type: object
      properties:
        entryType:
          type: string
          enum: [period_start, period_end, ovulation, symptom, note]
        date:
          type: string
          format: date
        symptoms:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - entryType
        - date
...
    DailyContext:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        sleepHours:
          type: number
          nullable: true
        energyLevel:
          type: integer
          nullable: true
        mood:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - createdAt
    CreateDailyContextBody:
      type: object
      properties:
        date:
          type: string
          format: date
        sleepHours:
          type: number
          nullable: true
        energyLevel:
          type: integer
          nullable: true
        mood:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - date
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error