# lib/api-zod/src/generated/types/createCycleEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Her Planner API - AI-powered life planner for women
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCycleEntryBodyEntryType } from "./createCycleEntryBodyEntryType";

export interface CreateCycleEntryBody {
  entryType: CreateCycleEntryBodyEntryType;
  date: Date;
  symptoms?: string | null;
  notes?: string | null;
}