# lib/api-zod/src/generated/types/cycleEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Her Planner API - AI-powered life planner for women
 * OpenAPI spec version: 0.1.0
 */
import type { CycleEntryEntryType } from "./cycleEntryEntryType";

export interface CycleEntry {
  id: number;
  entryType: CycleEntryEntryType;
  date: Date;
  symptoms?: string | null;
  notes?: string | null;
  createdAt: Date;
}