# lib/api-zod/src/generated/types/openaiError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Her Planner API - AI-powered life planner for women
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiError {
  error: string;
}